requirements.txt.txt
python-telegram-bot==22.3
yt-dlp
